third_party/WebKit/LayoutTests/fast/dom/htmlformcontrolscollection-enumerated-properties-expected.txt

This tests verifies the enumerated properties on HTMLFormControlsCollection and their order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS htmlFormControlsCollection.__proto__ is HTMLFormControlsCollection.prototype
PASS htmlFormControlsCollection.__proto__.__proto__ is HTMLCollection.prototype
PASS htmlFormControlsCollection.length is 6
PASS enumeratedProperties is expectedEnumeratedProperties
PASS successfullyParsed is true

TEST COMPLETE
     
